backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
openai>=1.0.0

# Database
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Celery & Redis
celery>=5.3.0
redis>=5.0.0
flower>=2.0.0

